fix API and Database relationship
