+++ utils/mongoDB/seeders/seedLetter.js
@@ -1,6 +1,6 @@
 require('dotenv').config();
-const Letter = require('../models/Letter.model');
-const { dbConnect } = require('../dbConnect');
+const Letter = require('../models/letter');
+const { dbConnect } = require('../MongoConnect');
 
 const seedLetterData =
     {
@@ -10,18 +10,21 @@ const seedLetterData =
 
 async function seedLetter() {
     try {
+        //database connection
         await dbConnect(); 
 
-        //check if table is already existed and drop if existed
+        //Validation
         const existedLetter = await Letter.find({});
         if(existedLetter.length !== 0) {
             await Letter.collection.drop();
             console.log("Existing Letter table dropped");
         }
-        //initial and save letter into database
+        //database logic
+        
         var savedLetter = new Letter(seedLetterData);
         savedLetter.save();
         console.log('Letter data seeded success');
+
     } catch (error) {
         console.log("Error while seeding data", error);
     }
